+++ blogs/db_router.js
@@ -34,4 +34,14 @@ router.get("/:id", async (req, res) => {
   }
 });
 
+router.post("/", async (req, res) => {
+  try {
+    const blog = await db.insert(req.body);
+    res.status(201).json(blog);
+  } catch (error) {
+    console.log(error);
+    res.status(500).json({ ErrorMessage: "Error adding blog" });
+  }
+});
+
 module.exports = router;
